requirements.txt
Flask==3.1.0
Flask_Cors==5.0.0
numpy==2.1.3
oct2py==5.7.2
Pillow==11.0.0
python-dotenv==1.0.1

